Split the two files past the cap and settle the gate findings
