+++ backend/internal/compose/attention/classify.go
@@ -419,64 +419,6 @@ func classifyUndelivered(item crmcontracts.AttentionItem, asOf time.Time) ranked
 	return ranked{item: row, occurredAt: occurredOf(item, asOf)}
 }
 
-// classifyDecay: a relationship going quiet. Nobody is waiting on the reader
-// for it, which is exactly why it goes unnoticed — and why it sits low rather
-// than not at all.
-//
-// How low depends on what the relationship was WORTH. One that still carries
-// money, or that was real before it went silent, is agreed work; a weak one
-// with nothing on it is routine. The same shape classifyWaiting uses for a
-// stale wait, and for the same reason: a lane where every row ranks alike is
-// one a rep reads once, and the rep's strongest contact going silent is not
-// the same morning's work as a cc drifting.
-func classifyDecay(item crmcontracts.AttentionItem, asOf time.Time) ranked {
-	level := levelRoutine
-	if decayMatters(item.Relationship) {
-		level = levelAgreed
-	}
-	row := base(item, level, "system", "data_drifts")
-	quiet := quietDaysOf(item)
-	if quiet > 0 {
-		row.Because = append(row.Because, reason("quiet_days", daysValue(quiet)))
-	}
-	// Why this silence outranks the others, in the vocabulary the contract
-	// already declares. Only the deal says so out loud: `expected_revenue`
-	// reads as "money rests on this" with no figure, which is exactly the claim.
-	//
-	// The BAND deliberately states nothing. It moves the level, and the reason
-	// vocabulary has no word for it — `no_champion` is the nearest and means
-	// the opposite here, an account with nobody carrying it rather than a
-	// person who WAS carrying it. A reason that reads backwards is worse than
-	// a rank the row does not explain, so the band waits for a word of its own.
-	if facts := item.Relationship; facts != nil && facts.HasOpenDeal != nil && *facts.HasOpenDeal {
-		row.Because = append(row.Because, reason("expected_revenue", nil))
-	}
-	return ranked{item: row, waitingDays: quiet, occurredAt: occurredOf(item, asOf)}
-}
-
-// decayMatters says whether a lapsed relationship is worth more than routine
-// attention: money still resting on it, or a real relationship behind it.
-//
-// A lane that sent no facts answers false rather than true. Absent is not a
-// claim, and promoting every silence on a field nobody filled would empty the
-// routine band for a reason no reader could see.
-func decayMatters(facts *crmcontracts.AttentionRelationshipFacts) bool {
-	if facts == nil {
-		return false
-	}
-	if facts.HasOpenDeal != nil && *facts.HasOpenDeal {
-		return true
-	}
-	if facts.Strength == nil {
-		return false
-	}
-	// Moderate counts. The §4 threshold for it already means a real two-way
-	// history, and a rep losing those quietly is who this lane is for —
-	// admitting only `strong` would leave it ranking almost nothing.
-	return *facts.Strength == crmcontracts.AttentionRelationshipFactsStrengthModerate ||
-		*facts.Strength == crmcontracts.AttentionRelationshipFactsStrengthStrong
-}
-
 // classifySystem: the pipes. A mailbox that stopped capturing makes every quiet
 // claim on this page suspect, which is why it is here at all rather than only in
 // an admin screen.
